refactor: rename formatter files
