+++ packages/graphback-codegen-resolvers/src/plugin/ResolverGeneratorPlugin.ts
@@ -1,6 +1,6 @@
 import { GraphbackCoreMetadata, GraphbackPlugin, ModelDefinition } from '@graphback/core';
 import { GraphQLSchema } from 'graphql';
-import { createRootResolversIndex } from '../formatters/apollo/apolloResolverFormatter';
+import { createRootResolversIndex } from '../formatters/apollo';
 import { generateCRUDResolvers, generateCustomCRUDResolvers } from './createResolvers';
 import { createCustomOutputResolvers, createOutputResolvers, OutputResolvers } from './outputResolvers';
 import { writeResolvers } from './writeResolvers';
